Move diff editor internal commands to their own file
